rename AuthForm to FormAuth
